feat: Backend - Minio, DTO, Repository

+++ springboot-app/src/main/java/com/lumos/lumosspring/minio/service/MinioService.kt
@@ -49,7 +49,7 @@ class MinioService(private val minioClient: MinioClient) {
         )
     }
 
-    fun getPresignedObjectUrl(bucketName: String, objectName: String, expiryAt: Long = 5 * 60): String {
+    fun getPresignedObjectUrl(bucketName: String, objectName: String, expiryAt: Int = 5 * 60): String {
         val url = minioClient.getPresignedObjectUrl(
             GetPresignedObjectUrlArgs.builder()
                 .method(Method.GET)
